important bugfix to path planning algorithm, which is also restructured.
The time input to the function mainm()/getPath() was set to t=1, and would not work for other time intervals. Now it accepts any time, but this is not given to it yet. File path_planner_test.cpp contains a testcase for the changed functions. Pay attention to functions getPath() and the functions above it in the file, the other functions were not changed in behavior.